requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
flower==1.2.0
humanize==4.4.0
kombu==5.2.4
packaging==21.3
prometheus-client==0.15.0
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.4
redis==4.3.4
six==1.16.0
tornado==6.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
